+++ tests/PSDocs.Azure.Tests/Azure.Conventions.Tests.ps1
@@ -31,10 +31,7 @@ Describe 'Conventions' -Tag 'Conventions' {
 
             # Generates templates
             $templatePath = Join-Path -Path $rootPath -ChildPath 'templates/';
-            $result = Get-AzDocTemplateFile -Path $templatePath | ForEach-Object {
-                $template = Get-Item -Path $_.TemplateFile;
-                Invoke-PSDocument @invokeParams -OutputPath $outputPath -InputObject $template.FullName -Convention 'Azure.NameByParentPath'
-            }
+            $result = Invoke-PSDocument @invokeParams -OutputPath $outputPath -InputPath $templatePath -Convention 'Azure.NameByParentPath'
             $result | Should -Not -BeNullOrEmpty;
             $filteredResult = $result | Where-Object { $_.Name -like 'acr_*' };
             $filteredResult.Name | Should -BeLike 'acr_v1.md';
@@ -46,10 +43,7 @@ Describe 'Conventions' -Tag 'Conventions' {
 
             # Generates templates without version path
             $templatePath = Join-Path -Path $here -ChildPath 'template-test/';
-            $result = Get-AzDocTemplateFile -Path $templatePath | ForEach-Object {
-                $template = Get-Item -Path $_.TemplateFile;
-                Invoke-PSDocument @invokeParams -OutputPath $outputPath -InputObject $template.FullName -Convention 'Azure.NameByParentPath'
-            }
+            $result = Invoke-PSDocument @invokeParams -OutputPath $outputPath -InputPath $templatePath -Convention 'Azure.NameByParentPath'
             $result | Should -Not -BeNullOrEmpty;
             $result[0].Name | Should -BeLike 'template-test.md';
             $result;